+++ narrative.rb
@@ -360,10 +360,14 @@ file 'app/forms/application_form.rb', <<~RUBY
       def attribute_names           = portrayal.keywords.without(:action, :method)
       def from_params(params, **kw) = new(**filter_params(params), **kw)
 
+      def public_attribute_names
+        attribute_names.select { |n| public_method_defined?(n) }
+      end
+
       def filter_params(params)
         params
           .require(param_key)
-          .permit(*attribute_names)
+          .permit(*public_attribute_names)
           .to_hash
           .transform_keys(&:to_sym)
       end